+++ src/monagement/tools/get_fl_fl.rs
@@ -1,5 +1,5 @@
-pub fn get_fl_sl(size: u32, minimum_size: u32, second_level_count: u32) -> (u32, u32) {
-    let size = if (size as u32) < minimum_size {
+pub fn get_fl_sl(size: u64, minimum_size: u64, second_level_count: u64) -> (u64, u64) {
+    let size = if size < minimum_size {
         minimum_size as f32
     } else {
         size as f32
@@ -8,5 +8,5 @@ pub fn get_fl_sl(size: u32, minimum_size: u32, second_level_count: u32) -> (u32,
     let fl = size.log2() as u32;
     let fl_indexing = fl - 2;
     let sl_indexing = (size as i32 - 2i32.pow(fl)) / (2i32.pow(fl) / second_level_count as i32);
-    (fl_indexing, sl_indexing as u32)
+    (fl_indexing as u64, sl_indexing as u64)
 }
